[requests] code refactoring in multiple modules
